+++ app/src/models/App.ts
@@ -1,7 +1,7 @@
 import * as dotenv from "dotenv";
 
 import express from "express";
-import { router } from "./routes/router";
+import { router } from "../routes/router";
 import path from "path";
 
 const directory = path.resolve(process.cwd(), "../../.env");
